+++ src/frontend/web_application/src/modules/notification/index.js
@@ -1,3 +1,2 @@
 export { default as NotificationProvider } from './components/NotificationProvider';
-export { default as MessageNotifications } from './components/MessageNotifications';
-export { default as MessageNotifier } from './components/MessageNotifier';
+export { default as MessageNotificationHandler } from './components/MessageNotificationHandler';